Change shared config
