+++ examples/ios/swift/AppClipParent/Constants.swift
@@ -1,0 +1,4 @@
+// add copyright
+struct entitlementConstants {
+    static let groupId = "group.TEAM_ID.com.domain.APP_GROUP"
+}
